updating to latest ollama
(The seeds give different answers.)
