+++ tests/e2e/test_e2e_haverscript/test_images.txt
@@ -2,25 +2,26 @@
 > Describe this image, and speculate where it was taken.
 
 This is a panoramic photograph depicting a cityscape with notable landmarks.
-In the foreground, there's a dense cluster of buildings, suggesting a
-well-established urban area. Dominating the middle ground is an old castle
-structure, which appears to be part of a larger fortification due to its
-integration into a hillside. The architecture suggests this might be a
-European city, possibly in Scotland given the presence of a castle with a
-distinctive roofline that resembles many historical buildings found there.
+In the foreground, there's a wide street lined with buildings and shops. On
+either side of the street, you can see various architectural styles suggesting
+a mix of historical and modern construction.
 
-The sky above is overcast, hinting at recent rainfall, and the lighting on the
-scene indicates it's either early morning or late afternoon. The trees at the
-top of the image are leafless, which could mean the season is autumn or
-winter, although this also suggests a possible urban setting as evergreen
-trees are often found in city centers.
+In the middle ground, there's an open space that leads to a significant
+building in the distance, which appears to be a castle due to its fortified
+walls and design elements typical of medieval architecture. The castle is
+likely an important landmark within the city.
 
-The overall style of the photograph seems to have been taken with a fisheye
-lens or a similar wide-angle technique, given the distortion at the edges and
-the perspective it provides. This type of photography is often used for
-panoramic shots to capture vast scenes such as cityscapes.
+Further back, beyond the castle, you can see more of the city, with trees and
+other buildings indicating a densely populated area. There's also a notable
+feature in the background that looks like a bridge or an overpass, spanning
+across a body of water which could be a river or a loch.
 
-Without more context or distinctive features that are unique to specific
-locations, it's difficult to pinpoint the exact location from this image
-alone. However, the architectural style and the presence of an old castle
-suggest a historical European city.
+The sky is partly cloudy, suggesting that the weather might be changing or
+it's just a typical day with some clouds. The vegetation appears lush and
+well-maintained, indicating that this city values its green spaces.
+
+Based on these observations, it seems likely that this image was taken in
+Edinburgh, Scotland. The castle in the distance is unmistakably Edinburgh
+Castle, one of the most iconic landmarks in Edinburgh. The architecture, the
+layout of the streets, and the presence of the castle are all indicative of
+the city's rich history and its blend of historical and modern elements.
